refactor: 스택 뷰 만드는 코드 파일 분리
